Przeniesienie przeciąganej listy zmiennych do osobnego komponentu.
